+++ wrapper/src/api/providers/benchmark/benchmarkProvider/cluster/test/utility.ts
@@ -1,4 +1,4 @@
-import { compareArrays } from '../../../../../../tools/comparators';
+import { compareArrays } from '../../../../../tools/comparators';
 import { NodeID } from '../types';
 import {
   RelaxedClustering,
