+++ wrapper/src/api/providers/benchmark/benchmarkProvider/idsToRecords/noTable.ts
@@ -1,5 +1,5 @@
-import { DatasetId, ExperimentIntersection } from '../../../../../server/types';
-import { DatasetIDMapper } from '../../../../dataset/datasetProvider/util/idMapper';
+import { DatasetId, ExperimentIntersection } from '../../../../server/types';
+import { DatasetIDMapper } from '../../../dataset/datasetProvider/util/idMapper';
 import { NodeID } from '../cluster/types';
 
 export function idClustersToRecordClustersNoTable(
