+++ wrapper/src/api/providers/benchmark/benchmarkProvider/intersection/calculatePairs/1Negative.ts
@@ -1,4 +1,4 @@
-import { Cluster, ClusterID } from '../../helper/cluster/types';
+import { Cluster, ClusterID } from '../../cluster/types';
 import { SubclusterCache } from '../cache';
 import { CalculatePairs } from './base';
